rf classes - combining classes
